refactor: MainDetailViewController에서 세그먼트 컨트롤을 PopupInfoDescriptionCollectionReusableView로 분리
